Next Milestone has been added, LogStream DB Driver
The DB Driver is almost finished and fully implemented. Currently it
is fixed on MYSQL, but it will be easy to implement other database
types later.

Filtering for the DB Driver works 75%, Date filtering is missing only.
I will work on this tomorrow. Another good thing is that we use DB Field
mapping which means it will be easy to adapt other database schemas later.