Minor changes, fixes and addons. Pager is now fully working.
The pager is now fully working for the LogStreamDB, and as far as possible
also in the LogStreamFile. The pager works with the full filtering as well.
Due the nature of file reading, of course very limitied.
Only Forward paging is possible.

Switch Advanced Form to generate search strings and redirect to the index.php
In short, this makes it easier to copy and send links to others. And also
makes it easier to later create predefined custom links.