Implemented pager button support for disk logstream
The pager button logic has been enhanced and modified
for unified support of all logstream classes. The disk logstream
supports a back and last button now. The database logstream has
be changed to work with the new logic. Minor other things have
been changed as well.

Sorting is disabled for all logstream classes now, it is simple
not implemented yet. If sorting for a field is available will be
decided by the logstream classes later.