Added detail page, which is linked from the list view.
The detail page contains all available details from a syslog
message. Also paging is possible in the detail view.
Minor fixes had to be done to make the detail page possible.

Added german translation to phplogcon.

Fixed a few notice bugs in the function_db.php, which
isn't really used at the moment. But will later when the
user-ui is added.