Changed install.php handling, so it is not necessary to remove the script anymore.
To avoid problems with automatic package updating, we changed the installer handling.
From now, the install.php will be allowed to run only in two cases, first if the
config.php file is missing, and second if the config.php file is empty.

Also removed the install file remove warning.