Added new feature of configureable views which can be configured and selected by source
If no VIEW is defined in the options of the current selected source,
the default VIEW will be selected. If no default VIEW is configured, the default will
be set ot the SYSLOG View.

The old Column format will also automatically converted into the new list of views.
So in case you have no default view set in your configuration, but the columns variable,
your old column format becomes the default view.