Fixed a problem when adding a mysql logstream source.
Due a problem of how php mysql_connect function handels second conenctions,
the Verify of a newly added logstream source could cause the first database
connection to be overwritten. The following error is, that the wrong
database was used for the UserDB after the new logstream was verified.

Also added a check into the redirect function to STOP redirecting if an
error happened in the UserDB before.